Turn on a bunch of warnings
Turn on as many warnings as easily possible from:
https://github.com/lefticus/cppbestpractices/blob/e73393f25a85f83fed7399d8b65cb117d00b2231/02-Use_the_Tools_Available.md#L100

Tested:
ipmitool sensor list still works

+++ src/HwmonTempSensor.cpp
@@ -103,7 +103,7 @@ void HwmonTempSensor::handleResponse(const boost::system::error_code& err)
         {
             float nvalue = std::stof(response);
             nvalue /= sensorScaleFactor;
-            if (nvalue != value)
+            if (static_cast<double>(nvalue) != value)
             {
                 updateValue(nvalue);
             }
