+++ src/PSUSensor.cpp
@@ -97,7 +97,7 @@ void PSUSensor::handleResponse(const boost::system::error_code& err)
             float nvalue = std::stof(response);
             responseStream.clear();
             nvalue /= sensorFactor;
-            if (nvalue != value)
+            if (static_cast<double>(nvalue) != value)
             {
                 updateValue(nvalue);
             }
